+++ src/shogun2-core/src/main/java/de/terrestris/shogun2/model/wps/WpsParameter.java
@@ -1,16 +1,21 @@
 package de.terrestris.shogun2.model.wps;
 
+import java.util.HashSet;
+import java.util.Set;
+
+import javax.persistence.CollectionTable;
+import javax.persistence.Column;
+import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
+import javax.persistence.JoinColumn;
 
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonSubTypes.Type;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
 import de.terrestris.shogun2.model.PersistentObject;
@@ -20,15 +25,7 @@ import de.terrestris.shogun2.model.PersistentObject;
  */
 @Entity
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
-@JsonTypeInfo(
-		use = JsonTypeInfo.Id.NAME,
-		property = "classType"
-)
-@JsonSubTypes({
-		@Type(value = WpsPrimitive.class, name = "WpsPrimitive"),
-		@Type(value = WpsReference.class, name = "WpsReference"),
-		@Type(value = WpsProcessExecute.class, name = "WpsProcessExecute")
-})
+@JsonTypeInfo(use = JsonTypeInfo.Id.CLASS, include = JsonTypeInfo.As.EXISTING_PROPERTY, property = "classType")
 public abstract class WpsParameter extends PersistentObject {
 
 	/**
@@ -47,10 +44,33 @@ public abstract class WpsParameter extends PersistentObject {
 	private String valueType;
 
 	/**
+	 *
+	 */
+	private String displayName;
+
+	/**
+	 * A set of formats this {@link WpsParameter} supports, e.g. 'text/xml;
+	 * subtype=gml/3.1.1' or 'application/wkt' or 'xs:double'
+	 */
+	@ElementCollection
+	@CollectionTable(joinColumns = @JoinColumn(name = "WPSPARAM_ID"))
+	@Column(name = "FORMAT_NAME")
+	private Set<String> supportedFormats = new HashSet<String>();
+
+	/**
+	 * A set of geometry types this {@link WpsParameter} supports, e.g. 'geom',
+	 * 'point', 'line', 'poly', 'multipoint', 'multiline', 'multipoly'
+	 */
+	@ElementCollection
+	@CollectionTable(joinColumns = @JoinColumn(name = "WPSPARAM_ID"))
+	@Column(name = "GEOMTYPE_NAME")
+	private Set<String> supportedGeometryTypes = new HashSet<String>();
+
+	/**
 	 * Constructor
 	 */
 	public WpsParameter() {
-		this.classType = getClass().getSimpleName();
+		this.classType = getClass().getName();
 	}
 
 
@@ -60,7 +80,6 @@ public abstract class WpsParameter extends PersistentObject {
 	public String getClassType() {
 		return classType;
 	}
-
 
 	/**
 	 * @return the valueType
@@ -78,6 +97,54 @@ public abstract class WpsParameter extends PersistentObject {
 	}
 
 	/**
+	 * @return the displayName
+	 */
+	public String getDisplayName() {
+		return displayName;
+	}
+
+
+	/**
+	 * @param displayName the displayName to set
+	 */
+	public void setDisplayName(String displayName) {
+		this.displayName = displayName;
+	}
+
+
+	/**
+	 * @return the supportedFormats
+	 */
+	public Set<String> getSupportedFormats() {
+		return supportedFormats;
+	}
+
+
+	/**
+	 * @param supportedFormats the supportedFormats to set
+	 */
+	public void setSupportedFormats(Set<String> supportedFormats) {
+		this.supportedFormats = supportedFormats;
+	}
+
+
+	/**
+	 * @return the supportedGeometryTypes
+	 */
+	public Set<String> getSupportedGeometryTypes() {
+		return supportedGeometryTypes;
+	}
+
+
+	/**
+	 * @param supportedGeometryTypes the supportedGeometryTypes to set
+	 */
+	public void setSupportedGeometryTypes(Set<String> supportedGeometryTypes) {
+		this.supportedGeometryTypes = supportedGeometryTypes;
+	}
+
+
+	/**
 	 * @see java.lang.Object#hashCode()
 	 *
 	 * According to
@@ -89,6 +156,7 @@ public abstract class WpsParameter extends PersistentObject {
 		return new HashCodeBuilder(19, 47) // two randomly chosen prime numbers
 			.appendSuper(super.hashCode())
 			.append(getValueType())
+			.append(getDisplayName())
 			.toHashCode();
 	}
 
@@ -108,6 +176,7 @@ public abstract class WpsParameter extends PersistentObject {
 		return new EqualsBuilder()
 			.appendSuper(super.equals(other))
 			.append(getValueType(), other.getValueType())
+			.append(getDisplayName(), other.getDisplayName())
 			.isEquals();
 	}
 
@@ -119,6 +188,7 @@ public abstract class WpsParameter extends PersistentObject {
 		return new ToStringBuilder(this, ToStringStyle.MULTI_LINE_STYLE)
 			.appendSuper(super.toString())
 			.append("valueType", valueType)
+			.append("displayName", displayName)
 			.toString();
 	}
 }
